requirements.txt
Flask
Flask-SQLAlchemy
google-generativeai
python-dotenv
Werkzeug
rapidfuzz
Pillow
opencv-python
